requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.4.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
Markdown==3.8
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
